Fix loading times in server addition and sync dir addition windows (#90)
Previously we scanned the entire remote for the list of folders before showing the window to add a new sync directory (the folder listings are used for autocompletions). This proved to be quite a long-running task when the remote contains a bunch of folders though - now we just check the list of folders for the currently entered path.

A somewhat-similar issue was also present when adding new sync remotes: It was of my knowledge that we only did a simple listing of the remote's root to see if we could connect fine, but it appears that we scanned and downloaded every file instead (using rclone's `operations/size` RC command). This has been changed to rclone's `operations/stat` RC command instead), which is immensely less expensive to run.