Only sync files when they're changed locally/remotely (#164)
Previously we'd be syncing files in an endless loop, which wasn't great from a UI/UX perspective. This new implementation checks the local and remote files against those recorded since the last sync, and uses those to figure out if a directory sync needs to be done.

Currently we sync the entire directory pair when a change is detected, but this is still a fair amount better than the implementation we had previously. We'll fine-tune out the implementation at a later time.