Remove notice about Celeste being alpha software from README
Celeste's file synchronization algorithm has been fairly stable for a bit (albiet with a few *minor* bugs), so I think it'd be good to remove the alpha software notice from the README. The main reason I had it listed was to let people know that they're at a non-little chance of potentionally losing their files, but now that things are getting fairly stable it should be fine now.

Regarding when Celeste will be in a stable state, that will happen once Celeste releases 1.0.0.